+++ agentpluginapi/__init__.py
@@ -16,3 +16,18 @@ from .payload_result import PayloadResult
 from .ping_scan_data import PingScanData
 from .port_scan_data import PortScanData
 from .target_host import PortScanDataDict, TargetHost, TargetHostPorts
+from .dropper_execution_mode import DropperExecutionMode
+from .i_linux_agent_command_builder import (
+    LinuxDownloadMethod,
+    LinuxDownloadOptions,
+    LinuxRunOptions,
+    ILinuxAgentCommandBuilder,
+)
+from .i_windows_agent_command_builder import (
+    WindowsDownloadMethod,
+    WindowsDownloadOptions,
+    WindowsRunOptions,
+    WindowsShell,
+    IWindowsAgentCommandBuilder,
+)
+from .i_agent_command_builder_factory import IAgentCommandBuilderFactory